requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
watchdog>=3.0.0
google-cloud-bigquery>=3.13.0
db-dtypes>=1.2.0
requests>=2.31.0
